// codepulse.ui/src/app/features/blog-post/model/blog-post.model.ts
import { Category } from "../../category/models/category.model";

export interface BlogPost{
    id: string;
    title: string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    publishedDate: Date;
    author: string;
    isVisible: boolean;
    categories: Category[];
}